Move config parser to separate file
